Rename vscode-app -> extension-app file to match convention
