+++ extensions/ql-vscode/src/common/vscode/extension-app.ts
