+++ extensions/ql-vscode/test/vscode-tests/no-workspace/databases/db-panel-selection.test.ts
@@ -8,7 +8,7 @@ import { DbTreeDataProvider } from "../../../../src/databases/ui/db-tree-data-pr
 import { DbItemKind } from "../../../../src/databases/db-item";
 import type { DbTreeViewItem } from "../../../../src/databases/ui/db-tree-view-item";
 import { SELECTED_DB_ITEM_RESOURCE_URI } from "../../../../src/databases/ui/db-tree-view-item";
-import { ExtensionApp } from "../../../../src/common/vscode/vscode-app";
+import { ExtensionApp } from "../../../../src/common/vscode/extension-app";
 import { createMockExtensionContext } from "../../../factories/extension-context";
 import { createDbConfig } from "../../../factories/db-config-factories";
 import { setRemoteControllerRepo } from "../../../../src/config";
